requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
django-filter==25.1
djangorestframework==3.16.0
filelock==3.16.1
platformdirs==4.3.6
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.6
